adds typing
